+++ pycaret/clustering/functional.py
@@ -5,9 +5,9 @@ import pandas as pd
 from joblib.memory import Memory
 
 from pycaret.clustering.oop import ClusteringExperiment
-from pycaret.internal.utils import check_if_global_is_not_none
 from pycaret.loggers.base_logger import BaseLogger
 from pycaret.utils.constants import DATAFRAME_LIKE, SEQUENCE_LIKE
+from pycaret.utils.utils import check_if_global_is_not_none
 
 _EXPERIMENT_CLASS = ClusteringExperiment
 _CURRENT_EXPERIMENT: Optional[ClusteringExperiment] = None
