+++ pycaret/clustering/oop.py
@@ -13,7 +13,7 @@ from pycaret.internal.logging import get_logger
 from pycaret.internal.pycaret_experiment.unsupervised_experiment import (
     _UnsupervisedExperiment,
 )
-from pycaret.internal.pycaret_experiment.utils import MLUsecase
+from pycaret.utils.utils import MLUsecase
 
 LOGGER = get_logger()
