+++ pycaret/containers/base_container.py
@@ -5,7 +5,7 @@
 import inspect
 from typing import Any, Dict, Optional
 
-import pycaret.internal.utils
+import pycaret.utils.utils
 
 
 class BaseContainer:
@@ -61,10 +61,10 @@ class BaseContainer:
         self.active = True
 
     def get_class_name(self):
-        return pycaret.internal.utils.get_class_name(self.class_def)
+        return pycaret.utils.utils.get_class_name(self.class_def)
 
     def get_package_name(self):
-        return pycaret.internal.utils.get_package_name(self.class_def)
+        return pycaret.utils.utils.get_package_name(self.class_def)
 
     def get_dict(self, internal: bool = True) -> Dict[str, Any]:
         """
