+++ pycaret/containers/metrics/classification.py
@@ -17,7 +17,7 @@ from sklearn.metrics._scorer import _BaseScorer
 import pycaret.containers.base_container
 import pycaret.internal.metrics
 from pycaret.containers.metrics.base_metric import MetricContainer
-from pycaret.internal.utils import get_label_encoder
+from pycaret.utils.utils import get_label_encoder
 
 
 class ClassificationMetricContainer(MetricContainer):
