+++ pycaret/internal/cuml_wrappers.py
@@ -1,13 +1,10 @@
 import numpy as np
-from scipy import sparse
 from sklearn.linear_model._base import LinearClassifierMixin
 from sklearn.multiclass import OneVsRestClassifier
 from sklearn.preprocessing import LabelBinarizer
 from sklearn.utils import check_array, column_or_1d
-from sklearn.utils.metaestimators import if_delegate_has_method
 from sklearn.utils.validation import _deprecate_positional_args, check_X_y
 
-from pycaret.internal.utils import get_all_object_vars_and_properties, is_fit_var
 from pycaret.utils._dependencies import _check_soft_dependencies
 
 if _check_soft_dependencies("cuml", extra=None, severity="warning"):
