+++ pycaret/loggers/mlflow_logger.py
@@ -2,9 +2,9 @@ import secrets
 from copy import deepcopy
 
 import pycaret
-from pycaret.internal.utils import mlflow_remove_bad_chars
 from pycaret.loggers.base_logger import BaseLogger
 from pycaret.utils import __version__
+from pycaret.utils.utils import mlflow_remove_bad_chars
 
 try:
     import mlflow
