+++ pycaret/regression/functional.py
@@ -1,15 +1,14 @@
 import logging
 from typing import Any, Callable, Dict, List, Optional, Union
 
-import numpy as np
 import pandas as pd
 from joblib.memory import Memory
 
 from pycaret.internal.parallel.parallel_backend import ParallelBackend
-from pycaret.internal.utils import check_if_global_is_not_none
 from pycaret.loggers.base_logger import BaseLogger
 from pycaret.regression.oop import RegressionExperiment
 from pycaret.utils.constants import DATAFRAME_LIKE, SEQUENCE_LIKE, TARGET_LIKE
+from pycaret.utils.utils import check_if_global_is_not_none
 
 _EXPERIMENT_CLASS = RegressionExperiment
 _CURRENT_EXPERIMENT: Optional[RegressionExperiment] = None
