+++ pycaret/regression/oop.py
@@ -26,9 +26,10 @@ from pycaret.internal.preprocess.preprocessor import Preprocessor
 from pycaret.internal.pycaret_experiment.supervised_experiment import (
     _SupervisedExperiment,
 )
-from pycaret.internal.pycaret_experiment.utils import MLUsecase, highlight_setup
 from pycaret.loggers.base_logger import BaseLogger
+from pycaret.utils import highlight_setup
 from pycaret.utils.constants import DATAFRAME_LIKE, SEQUENCE_LIKE, TARGET_LIKE
+from pycaret.utils.utils import MLUsecase
 
 LOGGER = get_logger()
