+++ pycaret/utils/utils.py
@@ -1,6 +1,7 @@
 import functools
 import inspect
 from copy import deepcopy
+from enum import Enum, auto
 from typing import Any, Callable, Dict, List, Optional, Set, Tuple, Union
 
 import numpy as np
@@ -17,6 +18,32 @@ from pycaret.internal.validation import (
     supports_partial_fit,
 )
 from pycaret.utils._dependencies import _check_soft_dependencies
+
+
+class MLUsecase(Enum):
+    CLASSIFICATION = auto()
+    REGRESSION = auto()
+    CLUSTERING = auto()
+    ANOMALY = auto()
+    TIME_SERIES = auto()
+
+
+def get_ml_task(y):
+    c1 = y.dtype == "int64"
+    c2 = y.nunique() <= 20
+    c3 = y.dtype.name in ["object", "bool", "category"]
+    if (c1 & c2) | c3:
+        ml_usecase = MLUsecase.CLASSIFICATION
+    else:
+        ml_usecase = MLUsecase.REGRESSION
+    return ml_usecase
+
+
+def highlight_setup(column):
+    return [
+        "background-color: lightgreen" if v is True or v == "Yes" else ""
+        for v in column
+    ]
 
 
 def get_classification_task(y):
